requirements.txt
google-play-scraper
langdetect
groq
google-generativeai
streamlit
python-dotenv
pyyaml
pandas
